+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - verboselogs
   - libblas
     - Defaults to using `accelerate` on Arm-based macOS, and `mkl` on Intel-based macOS and Linux
+  - movingpandas
 
 - Explicit Pip Packages
   - lxml